refactor: PreferencesRepository 인터페이스 추가
테스트 코드에 넣을 fake 객체를 정의하기 위해 인터페이스를 정의하고, 기존의 repository가 인터페이스를 구현하도록 변경했다.
